Обновлен MediaInfoLib v26.05-38-g702c9b7fd.

+++ src/ExtLib/MediaInfo/MediaInfo/Audio/File_Aac_Main.cpp
@@ -390,9 +390,9 @@ extern string Aac_ChannelLayout_GetString(const Aac_OutputChannel* const OutputC
 
     // Build the string
     string Value;
-    for (int i=0; i< OutputChannels_Size; i++)
+    for (size_t i=0; i< OutputChannels_Size; i++)
     {
-        if (OutputChannels[i]<Aac_OutputChannelPosition_Size)
+        if ((size_t)OutputChannels[i]<Aac_OutputChannelPosition_Size)
             Value+=Aac_OutputChannelPosition[OutputChannels[i]];
         else
             Value+=Ztring::ToZtring(OutputChannels[i]).To_UTF8();
@@ -493,9 +493,9 @@ extern string Aac_ChannelMode_GetString(const Aac_OutputChannel* const OutputCha
     // Count
     int8u ChannelModes[Aac_ChannelMode_Max+1];
     memset(ChannelModes, 0, Aac_ChannelMode_Max+1);
-    for (int i=0; i<OutputChannels_Size; i++)
+    for (size_t i=0; i<OutputChannels_Size; i++)
     {
-        if (OutputChannels[i]>=Aac_OutputChannelPosition_Size)
+        if ((size_t)OutputChannels[i]>=Aac_OutputChannelPosition_Size)
             ChannelModes[Aac_ChannelMode_Max]++;
         else
             ChannelModes[Aac_ChannelMode[OutputChannels[i]]]++;
